modified run time to new design

+++ test/sum-50.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-  volatile double x;
+  volatile float x;
   for (x = 0.0; x < 10.0; x += 0.2);
   printf("%.20g\n", x);
 }
